+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:video_player/video_player.dart';
 import 'package:video_player_win/video_player_win.dart';
 
 void main() {
@@ -16,15 +17,25 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  WinVideoPlayerController? controller;
+  VideoPlayerController? controller;
 
   void reload() {
     controller?.dispose();
-    controller = WinVideoPlayerController.file(File("E:\\test_youtube.mp4"));
+    controller = VideoPlayerController.file(File("E:\\test_youtube.mp4"));
+    //controller = WinVideoPlayerController.file(File("E:\\test_youtube.mp4"));
+    //controller = VideoPlayerController.networkUrl(Uri.parse("https://media.w3.org/2010/05/sintel/trailer.mp4"));
+    //controller = WinVideoPlayerController.file(File("E:\\Downloads\\0.FDM\\sample-file-1.flac"));
+
     controller!.initialize().then((value) {
       if (controller!.value.isInitialized) {
         controller!.play();
         setState(() {});
+
+        controller!.addListener(() {
+          if (controller!.value.isCompleted) {
+            log("ui: player completed, pos=${controller!.value.position}");
+          }
+        });
       } else {
         log("video file load failed");
       }
@@ -54,16 +65,18 @@ class _MyAppState extends State<MyApp> {
           title: const Text('video_player_win example app'),
         ),
         body: Stack(children: [
-          WinVideoPlayer(controller!),
+          VideoPlayer(controller!),
           Positioned(
               bottom: 0,
               child: Column(children: [
-                ValueListenableBuilder<WinVideoPlayerValue>(
+                ValueListenableBuilder<VideoPlayerValue>(
                   valueListenable: controller!,
                   builder: ((context, value, child) {
                     int minute = value.position.inMinutes;
                     int second = value.position.inSeconds % 60;
-                    return Text("$minute:$second",
+                    String timeStr = "$minute:$second";
+                    if (value.isCompleted) timeStr = "$timeStr (completed)";
+                    return Text(timeStr,
                         style: Theme.of(context).textTheme.headline6!.copyWith(
                             color: Colors.white,
                             backgroundColor: Colors.black54));
@@ -80,9 +93,17 @@ class _MyAppState extends State<MyApp> {
                     child: const Text("Pause")),
                 ElevatedButton(
                     onPressed: () => controller?.seekTo(Duration(
-                        milliseconds: controller!.value.position.inMilliseconds +
-                            10 * 1000)),
+                        milliseconds:
+                            controller!.value.position.inMilliseconds +
+                                10 * 1000)),
                     child: const Text("Forward")),
+                ElevatedButton(
+                    onPressed: () {
+                      int ms = controller!.value.duration.inMilliseconds;
+                      var tt = Duration(milliseconds: ms - 1000);
+                      controller?.seekTo(tt);
+                    },
+                    child: const Text("End")),
               ])),
         ]),
       ),