+++ lib/video_player_win.dart
@@ -16,6 +16,7 @@ class WinVideoPlayerValue {
   final bool isInitialized;
   final bool isLooping;
   final bool isPlaying;
+  final bool isCompleted;
   final double playbackSpeed;
   final Duration position;
   final Size size;
@@ -37,6 +38,7 @@ class WinVideoPlayerValue {
     this.isPlaying = false,
     this.isLooping = false,
     this.isBuffering = false,
+    this.isCompleted = false,
     this.volume = 1.0,
     this.playbackSpeed = 1.0,
     //int rotationCorrection = 0,
@@ -50,6 +52,7 @@ class WinVideoPlayerValue {
     bool? isInitialized,
     bool? isLooping,
     bool? isPlaying,
+    bool? isCompleted,
     double? playbackSpeed,
     Duration? position,
     Size? size,
@@ -62,6 +65,7 @@ class WinVideoPlayerValue {
       isInitialized: isInitialized ?? this.isInitialized,
       isLooping: isLooping ?? this.isLooping,
       isPlaying: isPlaying ?? this.isPlaying,
+      isCompleted: isCompleted ?? this.isCompleted,
       playbackSpeed: playbackSpeed ?? this.playbackSpeed,
       position: position ?? this.position,
       size: size ?? this.size,
@@ -71,117 +75,148 @@ class WinVideoPlayerValue {
 }
 
 class WinVideoPlayerController extends ValueNotifier<WinVideoPlayerValue> {
+  late final bool _isBridgeMode; // true if used by 'video_player' package
   int textureId_ = -1;
   final String dataSource;
   late final WinDataSourceType dataSourceType;
   bool _isLooping = false;
 
-  final _eventStreamController = StreamController<VideoEvent>(); // used for flutter official "video_player" package
+  // used by flutter official "video_player" package
+  final _eventStreamController = StreamController<VideoEvent>();
+
   Stream<VideoEvent> get videoEventStream => _eventStreamController.stream;
 
-  Future<Duration?> get position => Future.value(value.position);
+  Future<Duration?> get position async {
+    var pos = await _getCurrentPosition();
+    return Duration(milliseconds: pos);
+  }
 
-  WinVideoPlayerController._(this.dataSource, this.dataSourceType) : super(WinVideoPlayerValue())  {
+  WinVideoPlayerController._(this.dataSource, this.dataSourceType,
+      {bool isBridgeMode = false})
+      : super(WinVideoPlayerValue()) {
     if (dataSourceType == WinDataSourceType.contentUri) {
-      throw UnsupportedError("VideoPlayerController.contentUri() not supported in Windows");
+      throw UnsupportedError(
+          "VideoPlayerController.contentUri() not supported in Windows");
     }
     if (dataSourceType == WinDataSourceType.asset) {
-      throw UnsupportedError("VideoPlayerController.asset() not implement yet.");
+      throw UnsupportedError(
+          "VideoPlayerController.asset() not implement yet.");
     }
 
-    _finalizer.attach(this, this, detach: this);
+    _isBridgeMode = isBridgeMode;
     //VideoPlayerWinPlatform.instance.registerPlayer(_textureId, this);
   }
-  static final Finalizer<WinVideoPlayerController> _finalizer = Finalizer((player) {
-    // TODO: this finalizer seems not called...
-    player.dispose();
-    VideoPlayerWinPlatform.instance.unregisterPlayer(player.textureId_);
+  static final Finalizer<int> _finalizer = Finalizer((textureId) {
+    log("[video_player_win] gc free a player that didn't dispose() yet !!!!!");
+    VideoPlayerWinPlatform.instance.unregisterPlayer(textureId);
+    VideoPlayerWinPlatform.instance.dispose(textureId);
   });
 
-  WinVideoPlayerController.file(File file) : this._(file.path, WinDataSourceType.file);
-  WinVideoPlayerController.network(String dataSource) : this._(dataSource, WinDataSourceType.network);
-  WinVideoPlayerController.asset(String dataSource, {String? package}) : this._(dataSource, WinDataSourceType.asset);
-  WinVideoPlayerController.contentUri(Uri contentUri) : this._("", WinDataSourceType.contentUri);
+  WinVideoPlayerController.file(File file, {bool isBridgeMode = false})
+      : this._(file.path, WinDataSourceType.file, isBridgeMode: isBridgeMode);
+  WinVideoPlayerController.network(String dataSource,
+      {bool isBridgeMode = false})
+      : this._(dataSource, WinDataSourceType.network,
+            isBridgeMode: isBridgeMode);
+  WinVideoPlayerController.asset(String dataSource, {String? package})
+      : this._(dataSource, WinDataSourceType.asset);
+  WinVideoPlayerController.contentUri(Uri contentUri)
+      : this._("", WinDataSourceType.contentUri);
 
-  int _lastSeekId = 0; // +1 every time when seekTo() called, to cancel all getCurrentPosition() called before seekTo()
-  void _cancelTrackingPosition() => _lastSeekId++;
-  void _startTrackingPosition(int seekId) async {
-    // update position value for every 1000 ms
-    int delay = 1000 - (value.position.inMilliseconds % 1000);
-    delay = delay ~/ value.playbackSpeed + 1;
-    await Future.delayed(Duration(milliseconds: delay));
+  Timer? _positionTimer;
+  void _cancelTrackingPosition() => _positionTimer?.cancel();
+  void _startTrackingPosition() async {
+    // NOTE: 'video_player' package already auto get position periodically,
+    // so do nothing if _isBridgeMode = true
+    if (_isBridgeMode) return;
 
-    // NOTE: DO NOT call getCurrentPosition() during seeking, because windows will return 0...
-    if (textureId_ == -1 || !value.isInitialized || !value.isPlaying || (seekId != _lastSeekId)) return;
+    _positionTimer?.cancel();
+    _positionTimer =
+        Timer.periodic(const Duration(milliseconds: 300), (Timer timer) async {
 
-    var pos = await _getCurrentPosition();
-    if (pos < 0 || textureId_ == -1 || !value.isInitialized || (seekId != _lastSeekId)) return;
-    value = value.copyWith(position: Duration(milliseconds: pos));
+      if (!value.isInitialized
+          || !value.isPlaying
+          || value.isCompleted
+          || value.hasError) {
+        timer.cancel();
+        return;
+      }
 
-    _startTrackingPosition(seekId);
+      //log("[video_player_win] ui: position timer tick");
+      await position; // set player's position to value.position
+    });
   }
 
   void onPlaybackEvent_(int state) {
-    switch (state) { // MediaEventType in win32 api
+    switch (state) {
+      // MediaEventType in win32 api
       case 1: // MEBufferingStarted
         log("[video_player_win] playback event: buffering start");
         value = value.copyWith(isInitialized: true, isBuffering: true);
-        _eventStreamController.add(VideoEvent(eventType: VideoEventType.bufferingStart));
+        _eventStreamController
+            .add(VideoEvent(eventType: VideoEventType.bufferingStart));
         break;
       case 2: // MEBufferingStopped
         log("[video_player_win] playback event: buffering finish");
         value = value.copyWith(isInitialized: true, isBuffering: false);
-        _eventStreamController.add(VideoEvent(eventType: VideoEventType.bufferingEnd));
+        _eventStreamController
+            .add(VideoEvent(eventType: VideoEventType.bufferingEnd));
         break;
       case 3: // MESessionStarted , occurs when user call play() or seekTo() in playing mode
         //log("[video_player_win] playback event: playing");
-        value = value.copyWith(isInitialized: true, isPlaying: true);
-        _startTrackingPosition(_lastSeekId);
+        value = value.copyWith(
+            isInitialized: true, isPlaying: true, isCompleted: false);
+        _startTrackingPosition();
+        _eventStreamController
+            .add(VideoEvent(eventType: VideoEventType.isPlayingStateUpdate));
         break;
       case 4: // MESessionPaused
         //log("[video_player_win] playback event: paused");
         value = value.copyWith(isPlaying: false);
+        _cancelTrackingPosition();
+        _eventStreamController
+            .add(VideoEvent(eventType: VideoEventType.isPlayingStateUpdate));
         break;
       case 5: // MESessionStopped
         log("[video_player_win] playback event: stopped");
         value = value.copyWith(isPlaying: false);
+        _cancelTrackingPosition();
+        _eventStreamController
+            .add(VideoEvent(eventType: VideoEventType.isPlayingStateUpdate));
         break;
       case 6: // MESessionEnded
-        log("[video_player_win] playback event: play ended");
-        _cancelTrackingPosition();
-        value = value.copyWith(isPlaying: false, position: value.duration);
+        //log("[video_player_win] playback event: play ended");
         if (_isLooping) {
           seekTo(Duration.zero);
         } else {
-          // TODO: DO NOT send VideoEventType.completed to package [video_player]
-          // because [video_player] will call pause() then call seekTo(videoDuration)
-          // but seekTo in Windows cause play automatically
-          // since it start play again at the end, ,then we receive MESessionEnded again !
-          // and so on recurivly...
-          //_eventStreamController.add(VideoEvent(eventType: VideoEventType.completed));
+          _cancelTrackingPosition();
+          value = value.copyWith(isCompleted: true, position: value.duration);
+          _eventStreamController
+              .add(VideoEvent(eventType: VideoEventType.completed));
         }
         break;
       case 7: // MEError
         log("[video_player_win] playback event: error");
-        value = value.copyWith(isInitialized: false, hasError: true, isPlaying: false);
+        value = value.copyWith(
+            isInitialized: false, hasError: true, isPlaying: false);
+        _cancelTrackingPosition();
         break;
     }
   }
 
   Future<void> initialize() async {
-    WinVideoPlayerValue? pv = await VideoPlayerWinPlatform.instance.openVideo(this, textureId_, dataSource);
+    WinVideoPlayerValue? pv = await VideoPlayerWinPlatform.instance
+        .openVideo(this, textureId_, dataSource);
     if (pv == null) {
       log("[video_player_win] controller intialize (open video) failed");
       value = value.copyWith(hasError: true, isInitialized: false);
       _eventStreamController.add(VideoEvent(
-        eventType: VideoEventType.initialized,
-        duration: null,
-        size: null
-      ));
+          eventType: VideoEventType.initialized, duration: null, size: null));
       return;
     }
     textureId_ = pv.textureId;
     value = pv;
+    _finalizer.attach(this, textureId_, detach: this);
 
     _eventStreamController.add(VideoEvent(
       eventType: VideoEventType.initialized,
@@ -201,59 +236,23 @@ class WinVideoPlayerController extends ValueNotifier<WinVideoPlayerValue> {
     await VideoPlayerWinPlatform.instance.pause(textureId_);
   }
 
-  Timer? _seekTimer;
-  Completer? completer;
-  late Future<void> _lastSeekFuture;
-  int _lastSeekPos = -1;
-  Future<void> seekTo(Duration time) {
+  Future<void> seekTo(Duration time) async {
     if (!value.isInitialized) throw ArgumentError("video file not opened yet");
-    value = value.copyWith(position: time);
 
-    if (dataSourceType == WinDataSourceType.network) {
-      // for network source, we delay 300ms for each seekTo() call, and cancel last seekTo() call if next seekTo() called in 300ms
-      _lastSeekPos = time.inMilliseconds;
-      if (completer == null) {
-        _cancelTrackingPosition();
-        completer = Completer();
-        _lastSeekFuture = completer!.future;
-      }
-
-      _seekTimer?.cancel();
-      _seekTimer = Timer(const Duration(milliseconds: 300), () async {
-        _cancelTrackingPosition();
-        await VideoPlayerWinPlatform.instance.seekTo(textureId_, _lastSeekPos);
-        completer!.complete();
-        completer = null;
-      });
-      return _lastSeekFuture;
-    } else {
-      // for non-network source, we call seekTo() every 300ms
-      if (_seekTimer != null) {
-        _lastSeekPos = time.inMilliseconds;
-        return _lastSeekFuture;
-      } else {
-        completer = Completer();
-        _lastSeekFuture = completer!.future;
-        _seekTimer = Timer(const Duration(milliseconds: 300), () async {
-          // avoid user call seekTo() many times in a short time
-          _seekTimer = null;
-          if (_lastSeekPos >= 0) {
-            seekTo(Duration(milliseconds: _lastSeekPos)).then((_) {
-              completer!.complete();
-            });
-            _lastSeekPos = -1;
-          }
-        });
-
-        _cancelTrackingPosition();
-        return VideoPlayerWinPlatform.instance.seekTo(textureId_, time.inMilliseconds);
-      }
-    }
+    await VideoPlayerWinPlatform.instance
+        .seekTo(textureId_, time.inMilliseconds);
+    value = value.copyWith(position: time, isCompleted: false);
   }
 
   Future<int> _getCurrentPosition() async {
     if (!value.isInitialized) throw ArgumentError("video file not opened yet");
-    return await VideoPlayerWinPlatform.instance.getCurrentPosition(textureId_);
+    if (value.isCompleted) return value.duration.inMilliseconds;
+    int pos =
+        await VideoPlayerWinPlatform.instance.getCurrentPosition(textureId_);
+
+    if (textureId_ < 0) return 0;
+    value = value.copyWith(position: Duration(milliseconds: pos));
+    return pos;
   }
 
   Future<void> setPlaybackSpeed(double speed) async {
@@ -270,17 +269,19 @@ class WinVideoPlayerController extends ValueNotifier<WinVideoPlayerValue> {
 
   Future<void> setLooping(bool looping) async {
     _isLooping = looping;
+    value = value.copyWith(isLooping: looping);
   }
 
   @override
   Future<void> dispose() async {
     VideoPlayerWinPlatform.instance.unregisterPlayer(textureId_);
     await VideoPlayerWinPlatform.instance.dispose(textureId_);
+
+    _finalizer.detach(this);
     _cancelTrackingPosition();
 
     textureId_ = -1;
     value.textureId = -1;
-    value = value.copyWith(isInitialized: false, isPlaying: false);
     super.dispose();
 
     log("flutter: video player dispose: id=$textureId_");
@@ -292,8 +293,9 @@ class WinVideoPlayer extends StatefulWidget {
   final FilterQuality filterQuality;
 
   // ignore: unused_element
-  const WinVideoPlayer(this.controller, {Key? key, this.filterQuality = FilterQuality.low})
-    : super(key: key);
+  const WinVideoPlayer(this.controller,
+      {Key? key, this.filterQuality = FilterQuality.low})
+      : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _WinVideoPlayerState();
@@ -315,11 +317,11 @@ class _WinVideoPlayerState extends State<WinVideoPlayer> {
       color: Colors.black,
       child: Center(
         child: AspectRatio(
-            aspectRatio: widget.controller.value.aspectRatio,
-            child: Texture(
-              textureId: widget.controller.textureId_,
-              filterQuality: widget.filterQuality,
-            ),
+          aspectRatio: widget.controller.value.aspectRatio,
+          child: Texture(
+            textureId: widget.controller.textureId_,
+            filterQuality: widget.filterQuality,
+          ),
         ),
       ),
     );