+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 2.3.4
+
+* Fix: getCurrentPosition() shouldn't return 0 during seeking
+* support controller.value.isCompleted
+* auto call dispose() when controller garbage-collected
+
+## 2.3.2
+
+* Fix: sometimes crash when closing player
+
+## 2.3.1
+
+* Fix: cannot play video path containing non-ASCII characters
+
+## 2.3.0
+* Fix: get player position directly from windows media foundation, not cached value.
+* support Dart 3.0
+
 ## 2.2.2
 
 * Fix: shouldn't auto play when seeking in pause state